requirements_gui.txt
# Qt5 GUI dependencies
PyQt5>=5.15.0
PyQt5-Qt5>=5.15.0
PyQt5-sip>=12.8.0

# Original project dependencies (from requirements_slice.txt)
requests>=2.32.0
loguru>=0.7.0
pycryptodome>=3.20.0
ddddocr>=1.5.0
urllib3>=2.0.0